requirements.txt
django==2.2.6
django-cors-headers==3.1.1
django-rest-swagger==2.2.0
djangorestframework==3.9.4
